fix: product price to decimal type

+++ database/migrations/2022_03_08_125414_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('description');
-            $table->float('price');
+            $table->decimal('price', 4,2);
             $table->integer('stock');
         });
     }
